save lottie files
