+++ src/App.css
@@ -1,11 +1,11 @@
 #root {
-	max-width: 1280px;
-	margin: 0 auto;
-	padding: 2rem;
-	text-align: center;
+  max-width: 1280px;
+  margin: 0 auto;
+  padding: 2rem;
+  text-align: center;
 }
 
 nav {
-	border-bottom: "solid 1px";
-	padding-bottom: "1rem";
+  border-bottom: 'solid 1px';
+  padding-bottom: '1rem';
 }
