+++ src/components/Subtitle.tsx
@@ -1,7 +1,21 @@
-import React from 'react';
+import React, { useState } from 'react';
+import styled from 'styled-components';
 
 function Subtitle() {
-  return <div>Subtitle</div>;
+  return (
+    <SubtitleBox>
+      <p>Subtitle Subtitle Subtitle Subtitle Subtitle</p>
+    </SubtitleBox>
+  );
 }
+
+const SubtitleBox = styled.div`
+  position: sticky;
+  z-index: 100;
+  bottom: 2rem;
+  display: flex;
+  justify-content: center;
+  width: 100%;
+`;
 
 export default Subtitle;
